+++ RecipeApp/Recipe.swift
@@ -11,7 +11,7 @@ class Recipe {
     var title: String = ""
     var ingredients = [Ingredient]()
     var instructions:[String] = []
-    func addIngredient(newname: String, newquantity: Int, newmeasurement: String) {
+    func addIngredient(newname: String, newquantity: Float, newmeasurement: String) {
         ingredients.append(Ingredient(name: newname, quantity: newquantity, measurement: newmeasurement))
     }
     var numberOfInstructions = 0;
